Renames, file moves and cleanup
